Removed github head tags, refactored out Mocks
Merge added github tags in .sln file
Moved MockView and MockServiceLocator to their own files
