+++ Source/PrismLibrary_Tests.sln
@@ -29,11 +29,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prism.Mef.Wpf.Tests", "Wpf\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prism.Wpf.Tests", "Wpf\Prism.Wpf.Tests\Prism.Wpf.Tests.csproj", "{7DB8E74D-214F-4840-B294-58372D369958}"
 EndProject
-<<<<<<< HEAD
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prism.Ninject.Wpf", "Wpf\Prism.Ninject.Wpf\Prism.Ninject.Wpf.csproj", "{A4E80E69-EED3-440A-A50C-F64BBFC0103C}"
-=======
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prism.IocContainer.Wpf.Tests.Support", "Wpf\Prism.IocContainer.Wpf.Tests.Support\Prism.IocContainer.Wpf.Tests.Support.csproj", "{FF6F1B2C-BECC-4477-B918-B75615BAA3A5}"
->>>>>>> pr/19
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -85,17 +82,14 @@ Global
 		{7DB8E74D-214F-4840-B294-58372D369958}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DB8E74D-214F-4840-B294-58372D369958}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DB8E74D-214F-4840-B294-58372D369958}.Release|Any CPU.Build.0 = Release|Any CPU
-<<<<<<< HEAD
 		{A4E80E69-EED3-440A-A50C-F64BBFC0103C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A4E80E69-EED3-440A-A50C-F64BBFC0103C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4E80E69-EED3-440A-A50C-F64BBFC0103C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4E80E69-EED3-440A-A50C-F64BBFC0103C}.Release|Any CPU.Build.0 = Release|Any CPU
-=======
 		{FF6F1B2C-BECC-4477-B918-B75615BAA3A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FF6F1B2C-BECC-4477-B918-B75615BAA3A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF6F1B2C-BECC-4477-B918-B75615BAA3A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF6F1B2C-BECC-4477-B918-B75615BAA3A5}.Release|Any CPU.Build.0 = Release|Any CPU
->>>>>>> pr/19
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -110,10 +104,7 @@ Global
 		{F6443399-3E76-4B51-964A-95A166239D56} = {6BED0A37-BE6B-42D2-AAF0-C612FF57D9EB}
 		{B9BD4B23-8163-4EF3-955F-02B9578C45BA} = {6BED0A37-BE6B-42D2-AAF0-C612FF57D9EB}
 		{7DB8E74D-214F-4840-B294-58372D369958} = {6BED0A37-BE6B-42D2-AAF0-C612FF57D9EB}
-<<<<<<< HEAD
 		{A4E80E69-EED3-440A-A50C-F64BBFC0103C} = {6BED0A37-BE6B-42D2-AAF0-C612FF57D9EB}
-=======
 		{FF6F1B2C-BECC-4477-B918-B75615BAA3A5} = {6BED0A37-BE6B-42D2-AAF0-C612FF57D9EB}
->>>>>>> pr/19
 	EndGlobalSection
 EndGlobal
