+++ Source/Wpf/Prism.IocContainer.Wpf.Tests.Support/Mocks/MockView.cs
@@ -1,0 +1,6 @@
+﻿namespace Prism.IocContainer.Tests.Support.Mocks
+{
+    public class MockView
+    {
+    }
+}
