+++ Source/Wpf/Prism.Unity.Wpf.Tests/Mocks/MockServiceLocator.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Collections.Generic;
+using Microsoft.Practices.ServiceLocation;
+using Microsoft.Practices.Unity;
+using Prism.Regions;
+using Prism.Unity.Regions;
+
+namespace Prism.Unity.Tests.Mocks
+{
+    public class MockServiceLocator : ServiceLocatorImplBase
+    {
+        private readonly IUnityContainer _container;
+
+        public MockServiceLocator(IUnityContainer container)
+        {
+            _container = container;
+        }
+
+        protected override IEnumerable<object> DoGetAllInstances(Type serviceType)
+        {
+            throw new NotImplementedException();
+        }
+
+        protected override object DoGetInstance(Type serviceType, string key)
+        {
+            if (serviceType == typeof(IRegionNavigationService))
+            {
+                UnityRegionNavigationContentLoader loader = new UnityRegionNavigationContentLoader(this, _container);
+                return new RegionNavigationService(this, loader, new RegionNavigationJournal());
+            }
+
+            return null;
+        }
+    }
+}
