+++ Source/Wpf/Prism.Unity.Wpf.Tests/Prism.Unity.Wpf.Tests.csproj
@@ -8,7 +8,7 @@
     <ProjectGuid>{75511F4B-2602-4623-945C-AF8156B51F4E}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Prism.Unity.Wpf.Tests</RootNamespace>
+    <RootNamespace>Prism.Unity.Tests</RootNamespace>
     <AssemblyName>Prism.Unity.Wpf.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -100,6 +100,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Mocks\MockServiceLocator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UnityBootstrapperNullModuleManagerFixture.cs" />
     <Compile Include="UnityBootstrapperFixture.cs">
